refactor(package): add index file
affects: @fio/core, @fio/example
