+++ packages/core/index.ts
@@ -1,11 +1,11 @@
 /* tslint:disable:file-name-casing */
-export {Await} from './packages/core/src/main/Await'
-export {defaultRuntime} from './packages/core/src/runtimes/DefaultRuntime'
-export {FIO, IO, Task, TaskR, UIO} from './packages/core/src/main/FIO'
-export {FStream, Stream} from './packages/core/src/main/FStream'
-export {Fiber} from './packages/core/src/internals/Fiber'
-export {IRuntime} from './packages/core/src/runtimes/IRuntime'
-export {Managed} from './packages/core/src/main/Managed'
-export {Queue} from './packages/core/src/main/Queue'
-export {Ref} from './packages/core/src/main/Ref'
-export {testRuntime} from './packages/core/src/runtimes/TestRuntime'
+export {Await} from './src/main/Await'
+export {defaultRuntime} from './src/runtimes/DefaultRuntime'
+export {FIO, IO, Task, TaskR, UIO} from './src/main/FIO'
+export {FStream, Stream} from './src/main/FStream'
+export {Fiber} from './src/internals/Fiber'
+export {IRuntime} from './src/runtimes/IRuntime'
+export {Managed} from './src/main/Managed'
+export {Queue} from './src/main/Queue'
+export {Ref} from './src/main/Ref'
+export {testRuntime} from './src/runtimes/TestRuntime'
